AbRequest half days

+++ app/src/main/java/com/uthoff/dcm/android/repository/model/RequestDays.kt
@@ -1,12 +1,12 @@
 package com.uthoff.dcm.android.repository.model
 
 data class RequestDays(
-    var prevYear: Int,
-    var thisYear: Int,
-    var total: Int,
-    var corrections: Int,
-    var taken: Int,
-    var open: Int,
-    var thisRequest: Int,
-    var remaining: Int
+    var prevYear: Double,
+    var thisYear: Double,
+    var total: Double,
+    var corrections: Double,
+    var taken: Double,
+    var open: Double,
+    var thisRequest: Double,
+    var remaining: Double
 )
